Enhance the struct of code
Adjust some structures and include of our codes to make calculation
and configuration methods in right files.

* Move all the functions like limit() into a new file, mca_config.h.
* Rename mca.cpp with mca_config.cpp
* Update the #include "mca.h" in matrix.h with #include "mca_config.h"
* Add #include "mca_config.h" in other files when the files use functions
  like limit()

See #122
