+++ src/include/mca/mca.h
@@ -16,42 +16,6 @@
 
 namespace mca {
 using size_type = std::size_t;
-/* Initialize the mca, before using mca, you must call init
- * If you don't call init, mca will run in single thread mode
- * threadNum is how many threads will be used when calculating
- * limit is the minimal quantity of a thread's calculation
- * eps is the epsilon when comparing matrices whose elements' types are floating number */
-extern void init(const size_type &threadNum = std::thread::hardware_concurrency() - 1,
-                 const size_type &limit     = 623,
-                 const double &eps          = 1e-100);
-
-/* Set how many threads will be used when calculating */
-extern void setThreadNum(const size_type &threadNum);
-
-/* Set the minimal quantity of a thread's calculation
- * Make sure every sub-thread's calculation is no less than limit
- * When the rest part is less than limit, the main thread will calculate the rest
- * NOTE: if you want to calculate with single thread, you can set the limit with
- *       std::numeric_limits<size_type>::max() */
-extern void setLimit(const size_type &limit);
-
-/* Set the epsilon used for comparing floating numbers */
-extern void setEpsilon(const double &eps);
-
-/* Return current thread number */
-extern size_type threadNum();
-
-/* Return current limit */
-extern size_type limit();
-
-/* Return current epsilon */
-extern double epsilon();
-
-/* Return thread pool object, this should not called by the users, and this is for developers */
-extern ThreadPool &threadPool();
-
-/* Return calculation for every thread and the number of tasks */
-inline CalculationTaskNum threadCalculationTaskNum(const size_type &total);
 
 /* Check if matrix a and matrix b are equal using multi-thread
  * return false when a's shape is not same with b's shape
@@ -1074,13 +1038,6 @@ void transpose(const Matrix<T> &a, Matrix<O> &output) {
 
     // make sure all the sub threads are finished
     for (auto &item : returnValue) { item.get(); }
-}
-
-inline CalculationTaskNum threadCalculationTaskNum(const size_type &total) {
-    size_type calculation = std::max(total / (threadNum() + 1), limit());
-    size_type taskNum     = total / calculation;
-    if (total % calculation > 0) { taskNum++; }
-    return CalculationTaskNum{calculation, taskNum};
 }
 
 template <class T>
